move metrics to internal
